Fix some issues with the adventure game
Some typos are corrected. The exit door in the secret passage can now be
interacted with, it was getting pre-empted by the temple that also had
"door" as a synonym. Looking at the dragon now capitalizes more words,
giving more of an indication that the player is supposed to keep
looking.